Harden dictation godfiles across diff
