+++ src/app/components/workspace/composer/useComposerController.ts
@@ -7,19 +7,13 @@ import type {
   ComposerStreamingBehavior,
   ComposerThinkingLevel,
   DesktopActionInvoker,
-  DictationState,
 } from "../../../desktop/types";
 import { getDesktopActionErrorMessage } from "../../../desktop/action-results";
 import { useDismissibleLayer } from "../../../hooks/useDismissibleLayer";
-import {
-  appendDictatedText,
-  canUseLocalDictationCapture,
-  startLocalDictationCapture,
-  type LocalDictationCaptureSession,
-} from "./local-dictation";
 import { withComposerSendLock } from "./composerSendLock";
 import { mergeDraftWithRestoredQueuedPrompt } from "./composer-queue.helpers";
 import { composerDraftStore, getComposerDraftThreadId } from "./composerDraftStore";
+import { useComposerDictation } from "./useComposerDictation";
 import { submitComposerDraft } from "./submitComposerDraft";
 
 const thinkingLevelLabels: Record<ComposerThinkingLevel, string> = {
@@ -84,24 +78,22 @@ export function useComposerController({
   const [pendingPickerAttachments, setPendingPickerAttachments] = useState<ComposerAttachment[]>(
     [],
   );
+  const composerScopeKey = useMemo(
+    () => `${projectId}::${sessionPath ?? ""}::${draftThreadId ?? ""}`,
+    [draftThreadId, projectId, sessionPath],
+  );
   const pickerButtonRef = useRef<HTMLButtonElement>(null);
   const pickerPanelRef = useRef<HTMLDivElement>(null);
   const modelButtonRef = useRef<HTMLButtonElement>(null);
   const modelMenuRef = useRef<HTMLDivElement>(null);
+  const activeComposerScopeKeyRef = useRef(composerScopeKey);
   const activeDraftThreadIdRef = useRef<string | null>(draftThreadId);
   const skipNextDraftPersistenceRef = useRef<string | null>(null);
-  const dictationCaptureRef = useRef<LocalDictationCaptureSession | null>(null);
-  const [dictationActive, setDictationActive] = useState(false);
-  const [dictationInterimText, setDictationInterimText] = useState("");
-  const [dictationState, setDictationState] = useState<DictationState | null>(null);
   const draftValueRef = useRef("");
-  const dictationSessionTokenRef = useRef(0);
-  const dictationSettledPromiseRef = useRef<Promise<void> | null>(null);
-  const dictationSettledResolverRef = useRef<(() => void) | null>(null);
-  const dictationFlushPromiseRef = useRef<Promise<void> | null>(null);
   const sendLockRef = useRef(false);
 
   activeDraftThreadIdRef.current = draftThreadId;
+  activeComposerScopeKeyRef.current = composerScopeKey;
 
   const setDraftValue = useCallback((value: SetStateAction<string>) => {
     const nextValue =
@@ -111,133 +103,6 @@ export function useComposerController({
     draftValueRef.current = nextValue;
     setDraft(nextValue);
   }, []);
-
-  const resolveDictationSettled = useCallback(() => {
-    dictationSettledResolverRef.current?.();
-    dictationSettledResolverRef.current = null;
-    dictationSettledPromiseRef.current = null;
-  }, []);
-
-  const clearDictationSession = useCallback(() => {
-    dictationCaptureRef.current = null;
-    setDictationActive(false);
-    setDictationInterimText("");
-    resolveDictationSettled();
-  }, [resolveDictationSettled]);
-
-  const abortDictationSession = useCallback(() => {
-    dictationSessionTokenRef.current += 1;
-
-    const capture = dictationCaptureRef.current;
-    if (!capture) {
-      clearDictationSession();
-      return;
-    }
-
-    void capture.abort();
-    clearDictationSession();
-  }, [clearDictationSession]);
-
-  const stopDictationAndFlush = useCallback(async () => {
-    if (dictationFlushPromiseRef.current) {
-      await dictationFlushPromiseRef.current;
-      return;
-    }
-
-    const capture = dictationCaptureRef.current;
-    if (!capture) {
-      return;
-    }
-
-    dictationCaptureRef.current = null;
-
-    const submittedScopeKey = activeDictationScopeKeyRef.current;
-    const submittedSessionToken = dictationSessionTokenRef.current;
-    const flushPromise = (async () => {
-      setDictationActive(false);
-      setDictationInterimText("Transcribing…");
-
-      try {
-        const audio = await capture.stop();
-
-        if (
-          activeDictationScopeKeyRef.current !== submittedScopeKey ||
-          dictationSessionTokenRef.current !== submittedSessionToken
-        ) {
-          return;
-        }
-
-        if (!audio.audioBase64) {
-          setErrorMessage("No speech was captured.");
-          return;
-        }
-
-        if (!window.piDesktop?.transcribeDictation) {
-          setErrorMessage("Local dictation is unavailable in this runtime.");
-          return;
-        }
-
-        const result = await window.piDesktop.transcribeDictation({
-          audioBase64: audio.audioBase64,
-          sampleRate: audio.sampleRate,
-          language: navigator.language || null,
-        });
-
-        if (
-          activeDictationScopeKeyRef.current !== submittedScopeKey ||
-          dictationSessionTokenRef.current !== submittedSessionToken
-        ) {
-          return;
-        }
-
-        if (!result.ok) {
-          setErrorMessage(result.error ?? "Could not transcribe dictation.");
-          return;
-        }
-
-        if (result.text.trim()) {
-          setDraftValue((current) => appendDictatedText(current, result.text));
-          setErrorMessage(null);
-        }
-      } catch (error) {
-        if (
-          activeDictationScopeKeyRef.current === submittedScopeKey &&
-          dictationSessionTokenRef.current === submittedSessionToken
-        ) {
-          setErrorMessage(
-            error instanceof Error ? error.message : "Could not stop local dictation.",
-          );
-        }
-      } finally {
-        dictationFlushPromiseRef.current = null;
-
-        if (
-          activeDictationScopeKeyRef.current === submittedScopeKey &&
-          dictationSessionTokenRef.current === submittedSessionToken
-        ) {
-          clearDictationSession();
-        }
-      }
-    })();
-
-    dictationFlushPromiseRef.current = flushPromise;
-    await flushPromise;
-  }, [clearDictationSession, setDraftValue]);
-
-  const dictationScopeKey = useMemo(
-    () => `${projectId}::${sessionPath ?? ""}::${draftThreadId ?? ""}`,
-    [draftThreadId, projectId, sessionPath],
-  );
-  const activeDictationScopeKeyRef = useRef(dictationScopeKey);
-
-  activeDictationScopeKeyRef.current = dictationScopeKey;
-
-  useEffect(() => {
-    void dictationScopeKey;
-    abortDictationSession();
-  }, [abortDictationSession, dictationScopeKey]);
-
-  useEffect(() => abortDictationSession, [abortDictationSession]);
 
   useEffect(() => {
     skipNextDraftPersistenceRef.current = draftThreadId;
@@ -281,41 +146,23 @@ export function useComposerController({
   });
 
   const canSend = draft.trim().length > 0 && !isSending;
-  const dictationMissingModel = dictationState?.reason === "missing-model";
-  const dictationSupported = useMemo(
-    () =>
-      canUseLocalDictationCapture() &&
-      typeof window.piDesktop?.transcribeDictation === "function" &&
-      (dictationState?.available ?? true),
-    [dictationState],
-  );
 
-  useEffect(() => {
-    let disposed = false;
-
-    void dictationModelId;
-
-    if (!window.piDesktop?.getDictationState) {
-      return;
-    }
-
-    void window.piDesktop
-      .getDictationState()
-      .then((state) => {
-        if (!disposed) {
-          setDictationState(state);
-        }
-      })
-      .catch(() => {
-        if (!disposed) {
-          setDictationState(null);
-        }
-      });
-
-    return () => {
-      disposed = true;
-    };
-  }, [dictationModelId]);
+  const {
+    cancelDictation,
+    dictationActive,
+    dictationInterimText,
+    dictationMissingModel,
+    dictationSupported,
+    stopDictationAndFlush,
+    toggleDictation,
+  } = useComposerDictation({
+    dictationModelId,
+    draftThreadId,
+    projectId,
+    sessionPath,
+    setDraftValue,
+    setErrorMessage,
+  });
 
   const mergeAttachments = (current: ComposerAttachment[], next: ComposerAttachment[]) => {
     const byPath = new Map(current.map((attachment) => [attachment.path, attachment]));
@@ -367,7 +214,7 @@ export function useComposerController({
     }
 
     await withComposerSendLock(sendLockRef, async () => {
-      const submittedScopeKey = dictationScopeKey;
+      const submittedScopeKey = composerScopeKey;
       const submittedProjectId = projectId;
       const submittedSessionPath = sessionPath;
       const submittedDraftThreadId = draftThreadId;
@@ -378,7 +225,7 @@ export function useComposerController({
       try {
         await stopDictationAndFlush();
 
-        if (activeDictationScopeKeyRef.current !== submittedScopeKey) {
+        if (activeComposerScopeKeyRef.current !== submittedScopeKey) {
           return;
         }
 
@@ -456,71 +303,6 @@ export function useComposerController({
       setIsSending(false);
     }
   };
-
-  const toggleDictation = async () => {
-    if (dictationCaptureRef.current) {
-      void stopDictationAndFlush();
-      return "stopped" as const;
-    }
-
-    if (dictationFlushPromiseRef.current) {
-      await dictationFlushPromiseRef.current;
-      return "stopped" as const;
-    }
-
-    if (!canUseLocalDictationCapture() || !window.piDesktop?.transcribeDictation) {
-      setErrorMessage("Local dictation is unavailable in this runtime.");
-      return "unavailable" as const;
-    }
-
-    try {
-      const availability = window.piDesktop.getDictationState
-        ? await window.piDesktop.getDictationState().catch(() => null)
-        : dictationState;
-
-      if (availability) {
-        setDictationState(availability);
-      }
-
-      if (availability && !availability.available) {
-        if (availability.reason === "missing-model") {
-          setErrorMessage(null);
-          return "setup-required" as const;
-        }
-
-        setErrorMessage(availability.error ?? "Local dictation is unavailable in this runtime.");
-        return "unavailable" as const;
-      }
-
-      const capture = await startLocalDictationCapture();
-      dictationSessionTokenRef.current += 1;
-      dictationSettledPromiseRef.current = new Promise<void>((resolve) => {
-        dictationSettledResolverRef.current = resolve;
-      });
-      dictationCaptureRef.current = capture;
-      setDictationActive(true);
-      setDictationInterimText("");
-      setErrorMessage(null);
-      return "started" as const;
-    } catch (error) {
-      clearDictationSession();
-      setErrorMessage(error instanceof Error ? error.message : "Could not start local dictation.");
-      return "unavailable" as const;
-    }
-  };
-
-  const cancelDictation = useCallback(async () => {
-    if (dictationCaptureRef.current) {
-      abortDictationSession();
-      return;
-    }
-
-    if (dictationFlushPromiseRef.current) {
-      dictationSessionTokenRef.current += 1;
-      clearDictationSession();
-      await dictationFlushPromiseRef.current.catch(() => undefined);
-    }
-  }, [abortDictationSession, clearDictationSession]);
 
   const pickAttachments = async () => {
     if (openMenu === "picker") {
